serial: mvebu-uart: add TX interrupt trigger for pulse interrupts
Pulse interrupts (UART1 only) needs a change of state to trigger the TX
interrupt. In addition to enabling the TX_READY_INT_EN flag, produce a
FIFO state change from 'empty' to 'not full'. For this, write only one
data byte in TX start, making the TX FIFO not empty, and wait for the TX
interrupt to continue the transfer.

Signed-off-by: Allen Yan <yanwei@marvell.com>
Signed-off-by: Miquel Raynal <miquel.raynal@free-electrons.com>