mtd: nand: enable cached program operations
Cached program operations are already supported in the nand core but
were disabled until now for gain issue.

Gain seems to be fair now and gain is up to 45% throughput on
eraseblock write speed and up to 15% on 2 page write according to
mtd_speedtest module (tested with Macronix MX30LF2G18AC).

Signed-off-by: Miquel Raynal <miquel.raynal@free-electrons.com>