arm64: dts: marvell: armada-37xx: add second UART port
Add a node in Armada 37xx DTSI file for the second UART, with a
different compatible due to its extended IP which has some
differences with the first UART already in place.

Make use of this commit to also fully describe the first port and
use the same clear and named interrupt bindings for both ports.

UART0 uses level-interrupts. UART1 uses edge-triggered interrupts.

UART1 could use both level and edge-triggered interrupts by making use
of the Interrupt Controller in front of the GIC-500 but this option
should not be considered anymore because:
* This controller was originally dedicated to debug.
* Managing this controller registers involves several North-Bridge
  registers (that have actually nothing to do with North-Bridge) and
  would require an additionnal driver to be written.

Signed-off-by: Miquel Raynal <miquel.raynal@free-electrons.com>