mtd: nand: pxa3xx: enable controller with syscon property
Marvell recent boards like A7k/A8k do not boot with NAND flash
controller activated by default. Enabling the controller is a matter of
writing a single bit in a system controller register and is needed to be
bootloader independant.

Request the phandle node to the system controller and if it is present,
set the right bit.

Signed-off-by: Miquel Raynal <miquel.raynal@free-electrons.com>