mtd: nand: samsung: add datasheet minimum strength requirements for a chip
Samsung NAND chip K9F4G08U0D minimum ECC strength requirement is 1 bit
per 512 bytes. As the chip is not ONFI nor JEDEC and because of the lack
of these values, boards using it failed to probe the NAND controller
driver. Fix this by adding the default values if needed.

Signed-off-by: Miquel Raynal <miquel.raynal@free-electrons.com>