mtd: nand: remove useless fields from pxa3xx platform data
It was experimentally shown that every Marvell platform will need the
"enable arbiter" bit to be set, so there is no need for a separate
property.

The "num_cs" field is always 1 and for a good reason, the old driver
(pxa3xx_nand.c) could only handle one. The new driver that replaces it
(marvell_nand.c) can handle more, but better use device tree for such
description.

Signed-off-by: Miquel Raynal <miquel.raynal@free-electrons.com>