irqchip/irq-mvebu-icu: add support for System Error Interrupts (SEI)
An SEI driver provides an MSI domain through which it is possible to
raise SEIs.

Handle the NSR probe function in a more generic way to support other
type of interrupts (ie. the SEIs).

Each interrupt domain is a tree domain because of maximum 207 entries.
Reallocating an ICU slot is prevented by the use of an ICU-wide bitmap.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>