PCI: aardvark: Add support for the PCIe card warm reset pin
The Aardvark IP can trigger a reset signal upon hot reset or link
failure that will only reach the components on the board without
affecting the entire device (eg. only the endpoint card).

This line is multiplexed on MPPs so if it is not used as PCIe reset
and multiplexed for instance as a GPIO, the signals produced by the
PCIe IP won't affect the state of the line.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>