arm64: dts: marvell: Externalize PCIe macros from CP11x file
PCIe macros are specific to CP110 and will not fit CP115
constraints. To keep the same way the files are organized, just move
some macros out of the CP11x generic file and define them directly in
SoC DTSI, instead of defining single addresses in the SoC DTSI and
reusing them in macros.

In the end, the following SoC defines are dropped:
* CP11X_PCIE_IO_BASE
* CP11X_PCIE_MEM_BASE

While the following defines are moved out of the generic DT to be put
in the SoC files and replace the above definitions:
* CP11X_PCIEx_IO_BASE
* CP11X_PCIEx_MEM_BASE

As the CP11X_PCIEx_MEM_SIZE macro is also subject to change, we put it
in the SoC files as well.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>