lib/bch: Export the logic deriving the locator polynomial
It may be useful sometimes to just verify the validity of the
derivation of the locator polynomial without actually looking for an
accurate error detection.

With BCH algorithm, deriving the locator polynomial syndrome is pretty
fast while looking for the exact count and location of each error is
pretty time consuming.

Let's extract the part of bch_decode() which does this derivation so
that we can use it without spending too much time finding the
roots. The naming is not just "find the number of bitflips" because
this new function does not fully cover the error cases and finding the
real number of bitflips implies calling find_poly_roots() which is
long lasting.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>