lib/bch: Allow easy bit swapping
It seems that several hardware ECC engine use a swapped representation
of bytes compared to software. It means that when the software BCH
engine is working in conjunction with data generated with hardware, we
must swap the bits inside bytes, eg:

    0x0A = b0000_1010 -> b0101_0000 = 0x50

Make it possible just by flipping a boolean in the BCH control
structure.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>