mtd: rawnand: Move BBM flags to bbt_options
Bad block markers are tracked in the first, second or last page of a
bloc. This is not a 100% bad block table related thing but is also so
tied to bad block management that it deserves to be in
chip->bbt_options more than chip->options so move it to bbm.h (the
header name means "bad block management", not "bad block marker").

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>