mtd: rawnand: Move NAND_NO_BBM_QUIRK to bbt_options
This flag is controller dependent because most of the time the ECC
engine layout will be the responsible for not being able to preserve
the bad block markers. On the other side, it is only handled by the
bad block management code so I choose to move it there, with all the
other bad block related flags. In any case, it does not belong to
regular chip->options.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>