mtd: rawnand: Move subpage related flags to NAND controller flags
The subpage handling is a bit complex already but is "wrong". Fixing
it adds a bit more complexity, but this extra complexity will be
hidden by an upcoming change.

So, there is one situation (Samsung NAND manufacturer driver) where the
"no subpage write" flag is set by someone else than the NAND
controller itself. So the idea is to keep this flag as being a NAND
chip flag, but we will need to migrate any other user of this flag to
a controller flag.

For the "subpage read" flag, users are only NAND controller drivers so
there is no additional complexity as just renaming the field and
moving it to the NAND controller flags.

To clarify:
* NAND_NO_SUBPAGE_WRITE BIT is kept as a NAND chip option and only set
  by nand_samsung.c.
* NAND_CONTROLLER_NO_SUBPAGE_WRITE BIT is created as a NAND controller
  option and all other drivers than Samsung's are using it.
* NAND_SUBPAGE_READ is transformed as NAND_CONTROLLER_SUBPAGE_READ and
  moved to the controller flags.
* The macro is now: NAND_CONTROLLER_HAS_SUBPAGE_READS().

It is a mechanical change.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>