spi: mxic: Add support for pipelined ECC operations
Some SPI-NAND chips do not have a proper on-die ECC engine providing
error correction/detection. This is particularly an issue on embedded
devices with limited resources because all the computations must
happen in software, unless an external hardware engine is provided.

These external engines are new and can be of two categories: external
or pipelined. Macronix is providing both, the former being already
supported. The second, however, is very SoC implementation dependent
and must be instanciated by the SPI host controller itself.

An entire subsystem has been contributed to support these engines but
for this implementation is very SoC specific and thus two functions
cannot be part of the API: one to check if the ECC engine must be used
and the other to trigger transfers when needed. They are exported
through a specific header.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>