spi: spi-mem: Create a helper to gather all the supports_op checks
So far we check the support for:
- regular operations
- dtr operations
Soon, we will also need to check the support for ECC operations.

As the combinatorial will increase exponentially, let's gather all the
checks in a single generic function. This new helper will be called by
the exported functions, directly used by the different drivers.

Then, in a second time, we will add an ECC check and allow this new
helper to be directly used to avoid increasing dramatically the number
of new helpers needed to cover the {dtr-on/dtr-off, ecc-on/ecc-off}
situations, and perhaps others too. It will always be possible to create
abstraction helpers in the future if a particular combination is
regularly used.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>