spi: spi-mem: Export the spi_mem_generic_supports_op() helper
The combination of checks against the number of supported operations is
going to increase exponentially each time we add a new parameter. So far
we only had a dtr parameter. Now we are introducing an ECC parameter. We
need to make this helper available for drivers with specific needs,
instead of creating another set of helpers each time we want to check
something new. In the future if we see that many different drivers use
the same parameter values, we might be tempted to create a specific
helper for that. But for now, let's just make the generic one available.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>