net: mac802154: Inform device drivers about beacon operations
Let's create a couple of driver hooks in order to tell the device
drivers that beacons are being sent, allowing them to eventually apply
any needed configuration, or in the worst case to refuse the operation
if it is not supported. These hooks are optional and not implementing
them does not prevent the beacons operation to happen. Returning an
error from these implementations will however shut down the beacons
configuration entirely.

Co-developed-by: David Girault <david.girault@qorvo.com>
Signed-off-by: David Girault <david.girault@qorvo.com>
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>