net: ieee802154: Handle scanning complex channels
UWB PHYs work with complex channel which include a new parameter not yet
supported: the preamble codes. Let's offer the possibility to the MAC
layers to get the information about what preamble code is requested (if
any). If no preamble code mask is provided when scanning a complex
channel, the list of supported codes is retrieved and used instead.

The mean PRF can also be a useful parameter because in some cases it
will have an impact on the possible sub-channels to scan. If no mean PRF
is provided, the currently defined one will be used if any. Otherwise
the first available PRF is assumed (which will be considered the default
parameter).

Signed-off-by: Romuald Despres <romuald.despres@qorvo.com>
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>