net: ieee802154: Use a structure to carry the channel information
Right now a channel is only a {page, channel} couple which makes it easy
to pass as parameters when relevant. However this does not work anymore
when handling UWB channels. These PHYs use what we call "complex
channels" which also include a preamble code and a mean PRF. All these
information are needed to fully define a channel.

Let's change the way channels are defined by first introducing a simple
structure containing the two existing fields. Then we will create a
number of helpers to simplify common operations (such as comparing or
saving a channel) and finally we will add support for complex
channels. Doing this preparation step in the first place allows to
significantly reduce the diff lines when adding support for complex
channels which allow to focus on the actual insertions.

There is no functional change.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>