net: ieee802154: Support UWB complex channels
Bring support for preamble codes and mean PRFs.

A channel change on the UWB page should contain more information than
just the channel number: at least a preamble code.

The list of preamble codes supported depends on the UWB channel, the
supported PRFs and the DPS. We then add two additional capabilities: the
list of PRFs and if DPS is supported by the PHY transceiver.

Co-developed-by: Romuald Despres <romuald.despres@qorvo.com>
Signed-off-by: Romuald Despres <romuald.despres@qorvo.com>
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>