ieee802154: Allow the creation of coordinator interfaces on most devices
In order to be able to create coordinator interfaces, we need the
drivers to advertize that they support this type of interface. Fill in
the right bit in almost all the current ieee802154 device driver
capabilities to allow the creation of these coordinator interfaces.

Only the CA8210 which is a bit specific (hardMAC) cannot.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>