net: mac802154: The energy detection callback is not mandatory
Introduced a while ago, there is not a single user in-tree really
implementing this callback. It is likely used by out-of-tree drivers, so
the decision was to not drop it *now* as we need to further discuss it,
however there is no reason to keep this callback mandatory, so at least
drop it from the list of mandatory ops.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>