mtd: Fix an early error cleanup
add_mtd_device() is called when registering proper devices as well as
mtd partitions. When it is a partition, the mtd device has been added to
the parent's children list by the caller before the device gets
registered. Unfortunately, upon error, put_device() (which is the right
call in most cases) will trigger the mtd release partition hook which
loudly warns because the mtd device is still registered in a list (the
parent's partition list). In a regular remove path, this would not be
the case but as we are error-ing out, we need to avoid calling
put_device() *here*. It is safe to do so because all the cleanup logic
that put_device() would trigger is already handled by the caller in case
of error.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>