hexdump: Convert the ascii boolean into a flag variable
The print_hex_dump() prototype is already quite long and there are
already several hundred in-tree users. One argument is a boolean for
enabling the ascii output. This is one way to format the buffer. We
could think of other ways, which in this case might need other booleans
to be enabled. In order to avoid messing several times with the
prototype (and all the callers), let's switch to a 'flags'
variable which can be easily extended.

There is no behavioral change intended.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>
---
I tried to play with Coccinelle for the first time in order to address
the 600+ occurences of print_hex_dump, print_hex_dump_debug and
dynamic_hex_dump. I believe I managed to handle most of the cases
(compared using git grep) except several places enclosed inside
comments, '#if 0' and macros. I did not find an actual reason why the
kernel configuration would have an impact but apparently it did. I
didn't find advanced (and readable) resources on the Internet to fix
that, so here is the script. Nevertheless, it's a great and powerful
tool.

@@
expression E;
@@

(
print_hex_dump
|
print_hex_dump_debug
|
dynamic_hex_dump
)
 (E,...,
(
- true
+ DUMP_FLAG_ASCII
|
- 1
+ DUMP_FLAG_ASCII
|
- false
+ 0
)
 );